doge-queue/requirement.txt
google-api-python-client==1.8.2
flask==1.1.2
slackeventsapi==2.1.0
google-cloud-pubsub===1.4.2